LayoutTests/http/tests/storageAccess/request-storage-access-top-frame.html

<!DOCTYPE html>
<html>
<body onload="runTest()">
<script src="/js-test-resources/js-test.js"></script>
<script src="/js-test-resources/ui-helper.js"></script>
<script>
    description("Tests that top frame storage access is always granted in case the page requests it.");
    jsTestIsAsync = true;

    function activateElement(elementId) {
        var element = document.getElementById(elementId);
        var centerX = element.offsetLeft + element.offsetWidth / 2;
        var centerY = element.offsetTop + element.offsetHeight / 2;
        UIHelper.activateAt(centerX, centerY).then(
            function() { },
            function() {
                testFailed("Promise rejected.");
                finishJSTest();
            }
        );
    }

    var requestStorageAccessReturnValue;

    function makeRequestWithoutUserGesture() {
        document.requestStorageAccess();
        var promise = document.requestStorageAccess();
        promise.then(function (hasAccess) {
            requestStorageAccessReturnValue = hasAccess;
            shouldBe("requestStorageAccessReturnValue", "true");
            shouldBe("document.hasStorageAccess", "true");
            activateElement("theButton");
        });
    }

    function makeRequestWithUserGesture() {
        document.requestStorageAccess();
        var promise = document.requestStorageAccess();
        promise.then(function (hasAccess) {
            requestStorageAccessReturnValue = hasAccess;
            shouldBe("requestStorageAccessReturnValue", "true");
            shouldBe("document.hasStorageAccess", "true");
            finishJSTest();
        });
    }

    function runTest() {
        makeRequestWithoutUserGesture();
    }
</script>
<button id="theButton" onclick="makeRequestWithUserGesture()">Request Access</button>
</body>
</html>
