LayoutTests/http/tests/storageAccess/request-storage-access-top-frame-expected.txt

Tests that top frame storage access is always granted in case the page requests it.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS requestStorageAccessReturnValue is true
PASS document.hasStorageAccess is true
PASS requestStorageAccessReturnValue is true
PASS document.hasStorageAccess is true
PASS successfullyParsed is true

TEST COMPLETE
Request Access
